move bcos-test to testutils
