+++ test/unittests/libprotocol/FakeTransaction.h
@@ -21,8 +21,8 @@
 #include "interfaces/protocol/Exceptions.h"
 #include "libprotocol/protobuf/PBTransactionFactory.h"
 #include "libutilities/Common.h"
-#include "../libutils/HashImpl.h"
-#include "../libutils/SignatureImpl.h"
+#include "../../../testutils/HashImpl.h"
+#include "../../../testutils/SignatureImpl.h"
 #include <boost/test/unit_test.hpp>
 
 using namespace bcos;
