+++ test/unittests/libprotocol/PBTransactionReceiptTest.cpp
@@ -20,7 +20,7 @@
  */
 #include "FakeTransactionReceipt.h"
 #include "libprotocol/Common.h"
-#include "../libutils/TestPromptFixture.h"
+#include "../../../testutils/TestPromptFixture.h"
 
 using namespace bcos;
 using namespace bcos::protocol;
