+++ test/unittests/libtable/Entry.cpp
@@ -20,7 +20,7 @@
 #include "interfaces/storage/TableInterface.h"
 #include "libtable/Table.h"
 #include "libtable/TableFactory.h"
-#include "../libutils/TestPromptFixture.h"
+#include "../../../testutils/TestPromptFixture.h"
 #include <boost/test/unit_test.hpp>
 #include <iostream>
 #include <string>
