+++ test/unittests/libutilities/Base64Test.cpp
@@ -18,7 +18,7 @@
  */
 #include "libutilities/Base64.h"
 #include "libutilities/DataConvertUtility.h"
-#include "../libutils/TestPromptFixture.h"
+#include "../../../testutils/TestPromptFixture.h"
 #include <boost/test/unit_test.hpp>
 #include <iostream>
 #include <string>
