+++ test/unittests/libutilities/DataConvertUtilityTest.cpp
@@ -20,7 +20,7 @@
  */
 #include "libutilities/DataConvertUtility.h"
 #include "libutilities/Exceptions.h"
-#include "../libutils/TestPromptFixture.h"
+#include "../../../testutils/TestPromptFixture.h"
 #include <boost/test/unit_test.hpp>
 #include <cstdlib>
 #include <ctime>
