+++ test/unittests/main/main.cpp
@@ -18,14 +18,7 @@
  * @date 2021-02-24
  */
 #define BOOST_TEST_MODULE FISCO_BCOS_Tests
-#define BOOST_TEST_NO_MAIN
+#define BOOST_TEST_MAIN
 
 #include <boost/test/included/unit_test.hpp>
 #include <boost/test/unit_test.hpp>
-
-int main(int argc, const char* argv[])
-{
-    auto fakeInit = [](int, char**) -> boost::unit_test::test_suite* { return nullptr; };
-    int result = boost::unit_test::unit_test_main(fakeInit, argc, const_cast<char**>(argv));
-    return result;
-}
